test(craft): relocate wake-failure test to EDU, split docker posture tests
